rework integration test workflow
We were previously running integration tests twice: once on the
"current" version of Go and once on the previous version.  We weren't
testing with multiple TinyGo versions.

It's probably too much overhead to build and test 2x2 versions, so let's
just use the latest Go and TinyGo versions.  Now that we also build the
integration tests in the "Build Examples" action, we'll at least catch
build failures easily.