use SendAsync, not SendAsyncStreaming, for requests without a body
When SyndAsyncStreaming is called, it streams the body to the backend
using "Transfer-Encoding: chunked".  For requests with a body, this is
the behavior we want.  For requests without a body, we want to avoid
unnecessary chunked encoding, and have observed servers that error when
seeing it in certain contexts. Calling SendAsync instead causes the
entire body to be buffered (in this case, zero bytes) and
"Content-Length" to be set instead.