fix a handful of backend bugs
Replace `backend.IsHealthy` with `backend.Health`, because it returns an
enum of 3 states (unknown, healthy, unhealthy) and not a bool.

When populating the backend, the runtime will return `FastlyStatusNone`
if a value isn't set.  This isn't an error, so we should ignore it.
Similarly, the Viceroy environment will return `FastlyStatusUnsupported`
for some unsupported fields and we should ignore them too.

The exported names for getting SSL min/max versions were copypasta'd and
returned the wrong values.

Fixes #65.