don't defer closing the response writer
Closing the response writer signifies to the Compute runtime that the
response has been successfully completed, so we shouldn't defer it in
case the handler panics.

In TinyGo, defers aren't run after panic so this wasn't executed.  In
Go, defers are run.  The result of this was that if a handler panicked
before sending a response, it would send an empty 200 OK response in Go
but a 500 Internal Server Error in TinyGo.

Fixes #98.