add erl.Policy struct type
This encapsulates several arguments that were previously passed into
the RateLimiter.CheckRate method.  It also implements the
json.Unmarshaler interface.  This makes it easier to have multiple
policies to easily check against, and the schema for JSON unmarshaling
makes it easier to define policies in things like Config Store entries.