Remove internal non-collapsing lookup
We can't accurately track "origin miss" vs. "pop miss" without knowing
the backend. We want to encourage lookups to be completed (i.e.
transactional lookups) so we can provide accurate CHR metrics.

This function is not exposed outside of internal/, and is not used by
the public API; so, remove it.