Respect range during cache streaming (#169)
The core cache API allows for cache items to be concurrently streamed to / from the cache.

Previously, if:
- The size of the cached item's body was not provided by the writer
- The reader requests a specific range of the cached item's body
- The writer and reader are concurrent, i.e. the body is streamed from one to the other

then the core cache API will ignore the requested range and provide the whole body. There is no explicit notification that the whole body is provided instead of a range.

In this SDK release we change the default: body reads through the cache will use the requested range even when streaming. A lookup option allows setting the legacy behavior.