don't fire off goroutine if underlying body is *fastly.HTTPBody
We set streaming = false, copy data synchronously, and call SendAsync
for nil bodies or in-memory readers like bytes.Buffer, etc.  We should
also do that for *fastly.HTTPBody, since we're not even copying the
data, we're just signaling that the body is finished.  This prevents
us from starting up a goroutine and calling SendAsyncStreaming
unnecessarily.