make ResponseWriter.Close() idempotent, fix its docs
The docs were pretty strict about callers needing to call
ResponseWriter.Close(), but this wasn't true and would actually cause
hostcall errors.  The ResponseWriter is automatically closed after the
handler returns, and if user code did it first the runtime would always
throw InvalidHandle errors on it.

This loosens up the docs and tracks whether the response body has been
closed and doesn't try it again if so, eliminating the hostcall error.