add ServeManyOptionsFromContext, add serve-many e2e test
When fsthttp.ServeMany() is called, we now associate the options with
the context that gets passed to handlers.  ServeManyOptionsFromContext()
allows handlers to retrieve those options and modify them.

A new end-to-end test demonstrates this (and allows us to test the
reusable session functionality).

One possible use case for this is for Compute applications to monitor
their heap usage, and reset ServeManyOptions.MaxRequests to 1 if it
crosses some threshold to ensure that no more requests are handled by
that session.