handle invalid 1xx status codes in client responses
Compute does not allow 1xx status codes (except 103 Early Hints) and
returns FastlyStatus::Inval from SendDownstream if it is set.  We don't
check that call for errors and that causes us to silently return an
empty 200 OK response.

Because WriteHeader is infallible, we now capture the error from
SendDownstream on the responseWriter and return it from subsequent
Write, Append, and Close calls.  For the specific case of an invalid
status code, we translate it into a new exported ErrInvalidStatusCode
error.

Because calling SendDownstream consumes the underlying ABI response and
body even in failure, this situation is unrecoverable.  serve() checks
for that SendDownstream error and panics, ensuring that clients will
receive a 500 status code rather than a 200.

The docs for WriteHeader have been updated to explain that 1xx codes
are not permitted, and describe the error behavior.